fix : anyway-home directory remove * Main Componenets 생성
